Add localization file
