+++ lib/main.dart
@@ -1,77 +1,13 @@
 import 'package:flutter/material.dart';
-import 'package:google_maps_flutter/google_maps_flutter.dart';
-import 'screens/map.dart';
-import 'dart:async';
-import 'widgets/custom_appbar.dart';
-import 'package:google_fonts/google_fonts.dart';
 import 'package:flutter_localizations/flutter_localizations.dart';
-import 'package:intl/intl.dart';
-import 'l10n/messages_all.dart';
 import 'package:flutter_statusbarcolor/flutter_statusbarcolor.dart';
+import 'package:google_maps_flutter/google_maps_flutter.dart';
+import 'package:google_fonts/google_fonts.dart';
+import 'dart:async';
 
-//TODO: Move to new file.
-class ConcordiaLocalizations {
-  ConcordiaLocalizations(this.localeName);
-
-  static Future<ConcordiaLocalizations> load(Locale locale) {
-    final String name =
-        locale.countryCode.isEmpty ? locale.languageCode : locale.toString();
-    final String localeName = Intl.canonicalizedLocale(name);
-
-    return initializeMessages(localeName).then((_) {
-      return ConcordiaLocalizations(localeName);
-    });
-  }
-
-  static ConcordiaLocalizations of(BuildContext context) {
-    return Localizations.of<ConcordiaLocalizations>(
-        context, ConcordiaLocalizations);
-  }
-
-  final String localeName;
-
-  String get profile {
-    return Intl.message(
-      'Profile',
-      name: 'profile',
-      desc: 'Profile for the user',
-      locale: localeName,
-    );
-  }
-
-  String get schedule {
-    return Intl.message(
-      'Schedule',
-      name: 'schedule',
-      desc: 'Schedule',
-      locale: localeName,
-    );
-  }
-
-  String get settings {
-    return Intl.message(
-      'Settings',
-      name: 'settings',
-      desc: 'Settings',
-      locale: localeName,
-    );
-  }
-}
-
-class ConcordiaLocalizationsDelegate
-    extends LocalizationsDelegate<ConcordiaLocalizations> {
-  const ConcordiaLocalizationsDelegate();
-
-  @override
-  bool isSupported(Locale locale) => ['en', 'fr'].contains(locale.languageCode);
-
-  @override
-  Future<ConcordiaLocalizations> load(Locale locale) =>
-      ConcordiaLocalizations.load(locale);
-
-  @override
-  bool shouldReload(ConcordiaLocalizationsDelegate old) => false;
-}
+import 'screens/map.dart';
+import 'localization.dart';
+import 'widgets/custom_appbar.dart';
 
 void main() {
   runApp(App());
